backend-data/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
motor>=3.4.0
pymongo>=4.6.3
python-dotenv>=1.0.1
pydantic>=2.6.4
httpx>=0.27.0
